Refactor user model and router
This commit refactors the user model and router by replacing the
`getPublicProfile` method with `toJSON` method in the user model. This
change converts the user object to a plain JavaScript object and removes
the password field. The router is also updated to send the user object
directly instead of calling the `getPublicProfile` method.